requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
nltk==3.7
numpy==1.24.2
Pillow==9.4.0
pip==23.0.1
python-dotenv==1.0.0
requests==2.28.2
setuptools==65.6.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wheel==0.38.4
